+++ usecases/LDpred2_test/unittest/fun.R
@@ -3,6 +3,9 @@ dirTests <- Sys.getenv('DIR_SCRIPTS')
 source(paste0(dirTests, '/fun.R'))
 context("Test functions")
 dta <- data.frame(a=1:3, b=c(1, 'X', 2))
-test_that("Test numeric counts",
-          expect_equal(c(1, 3), getNumericIndices(dta$b))
-          )
+test_that("Test numeric counts", {
+  expect_true(isNumeric('1'))
+  expect_false(isNumeric('a'))
+  expect_true(isNumeric(0))
+  expect_equal(c(1, 3), getNumericIndices(dta$b))
+  })